+++ examples/practice-typescript/src/components/TableHeader/index.scss
